+++ CHANGELOG.md
@@ -1,2 +1,9 @@
+# 1.0.0
+* add unsigned little endian base 128 (ULEB128) packing and unpacking via BigInt
+* add unicode string packing/unpacking
+* add uint64 packing/unpacking
+* add BitStream.measure for measuring how many bits a stream would write to
+* add fuzz testing
+
 # 0.1.0
 * initial commit